fix : 코드 리뷰 적용 custom ui 삭제 modified 상태 삭제 호버 시 쿼리 미리보기 삭제 선택된 쿼리를 객체로 상태 저장하도록 변경
